+++ CHANGELOG.md
@@ -4,10 +4,41 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+
+## [0.2.1]
+### Added
+### Changed
+
+### Fixed
+- Fixed document links in README.md.
 
 
-## [0.1.3] - 2024-11-??
+
+## [0.2.0]
+### Added
+- Added ssl headers to nginx.conf (NginxConfig.generate/4)
+- Added `mix horizon.ops.init` to install HorizonOps scripts.
+- Added .ssh/environment file to store environment variables for ssh connections.
+- Added --verbose option on bsd_install.sh (similar to debug)
+- Added docs for Horizon Ops scripts.
+- Added docs for Proxmox setup.
+- Added docs for FreeBSD Install
+- Added docs for Hetzner Cloud setup
+- Added sample config files for hosts, e.g. [build, web, pg, backup]
+
+### Changed
+- Moved last two arguments of NginxConfig.send/4 to opts
+- Changed log_duration to `off` in postgresql.conf to save space.
+- Changed log_statement to `ddl` in postgresql.conf to save space.
+- Changed `mix horizon.init` to only install project specific deploy scripts.
+- Removed `-u user` option on ops scripts and included optional [user@] for hosts
+
+### Fixed
+- Fixed bug in NginxConfig.generate to properly handle multiple server names.
+
+
+
+## [0.1.3]
 ### Changed
 - Added `freebsd_setup.sh` script to help setup first-boot configuration for FreeBSD hosts.
 - Changed host on `bsd_install.sh` to combine user and host into a single argument.
@@ -15,6 +46,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Fixed
 - Fixed documentation issues.
+
+
 
 ## [0.1.2] - 2024-10-29
 ### Changed